+++ test/test-host.ts
@@ -5,14 +5,17 @@ import {
 } from "@typespec/compiler/testing";
 import type { EmitterOptions } from "../src/lib.js";
 
-const BaseTester = createTester(resolvePath(import.meta.dirname, "../.."), {
-  libraries: [
-    "@massivescale/tsp-ts-client-models",
-    "@typespec/http",
-    "@typespec/rest",
-    "@typespec/versioning",
-  ],
-});
+export const BaseTester = createTester(
+  resolvePath(import.meta.dirname, "../.."),
+  {
+    libraries: [
+      "@massivescale/tsp-ts-client-models",
+      "@typespec/http",
+      "@typespec/rest",
+      "@typespec/versioning",
+    ],
+  },
+);
 
 function testerFor(options?: EmitterOptions) {
   return BaseTester.emit(